fix: pinCard의 지도에서보기 버튼 기능 추가
